rope: move RopeSlice to its own file + lib.rs was getting too massive, so I thought this was a good idea + also, added some documentation
