+++ an-rope/src/lib.rs
@@ -28,8 +28,10 @@ use std::fmt;
 
 #[cfg(test)]
 mod test;
+mod slice;
 
 use self::internals::Node;
+pub use self::slice::{RopeSlice, RopeSliceMut};
 
 /// A Rope
 ///
@@ -642,33 +644,16 @@ impl Rope {
         self.bytes().zip(other).all(|(a, b)| a == b)
     }
 
+    #[inline]
     pub fn slice<'a, R>(&'a self, range: R) -> RopeSlice<'a>
     where R: RangeArgument<usize> {
-        let len = self.len();
+        RopeSlice::new(&self.root, range)
+    }
 
-        // if the RangeArgument doesn't have a defined start index,
-        // the slice begins at the 0th index.
-        let start = *range.start().unwrap_or(&0);
-        // similarly, if there's no defined end, then the end index
-        // is the last index in the Rope.
-        let end = *range.end().unwrap_or(&self.len());
-
-        let slice_len = end - start;
-
-        // find the lowest node that contains both the slice start index and
-        // the end index
-        let (node, offset) = if start == 0 && end == len {
-            // if the slice contains the entire rope, then the spanning node
-            // is the root node
-            (&self.root, 0)
-        } else {
-            self.root.spanning(start, slice_len)
-        };
-
-        RopeSlice { node: node
-                  , offset: offset
-                  , len: slice_len
-                  }
+    #[inline]
+    pub fn slice_mut<'a, R>(&'a mut self, range: R) -> RopeSliceMut<'a>
+    where R: RangeArgument<usize> {
+        RopeSliceMut::new(&mut self.root, range)
     }
 }
 
@@ -953,147 +938,5 @@ impl ops::IndexMut<ops::RangeTo<usize>> for Rope {
 impl ops::IndexMut<ops::RangeFrom<usize>> for Rope {
     fn index_mut(&mut self, i: ops::RangeFrom<usize>) -> &mut str {
         unimplemented!()
-    }
-}
-
-//-- rope slices ------------------------------------------------------------
-
-/// An immutable borrowed slice of a `Rope`.
-///
-/// A RopeSlice represents an immutable borrowed slice of some or all the
-/// characters in a `Rope`.
-#[derive(Debug)]
-pub struct RopeSlice<'a> { node: &'a Node
-                         , offset: usize
-                         , len: usize
-                         }
-
-impl<'a> fmt::Display for RopeSlice<'a> {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        // TODO: don't create an intermediate string?
-        write!(f, "{}", self.chars().collect::<String>())
-    }
-}
-
-/// An mutable borrowed slice of a `Rope`.
-///
-/// A `RopeSliceMut` represents a mutable borrowed slice of some or all the
-/// characters in a `Rope`.
-#[derive(Debug)]
-pub struct RopeSliceMut<'a> { node: &'a mut Node
-                            , offset: usize
-                            , len: usize
-                            }
-
-impl<'a> RopeSlice<'a> {
-    #[inline]
-    fn slice_char_iter<I, T>(&'a self, i: I) -> impl Iterator<Item=T> + 'a
-    where I: Iterator<Item=T>
-        , I: 'a
-        , T: Copy {
-            i.skip(self.offset).take(self.len)
-    }
-
-    fn slice_strings_iter<I>(&'a self, i: I) -> impl Iterator<Item=&'a str> +'a
-    where I: Iterator<Item=&'a str>
-        , I: 'a {
-        i.scan((self.offset, self.len), |curr, s| {
-            match *curr {
-                (0, 0) => None
-              , (0, ref mut remaining) if *remaining < s.len() => {
-                    let r = *remaining;
-                    *remaining = 0;
-                    Some(&s[..r])
-                }
-              , (0, ref mut remaining) => {
-                    *remaining -= s.len();
-                    Some(s)
-                }
-              , (ref mut offset, _) if *offset > s.len() => {
-                    *offset -= s.len();
-                    Some("")
-                }
-              , (ref mut offset, _) => {
-                    let c = *offset;
-                    *offset -= s.len();
-                    Some(&s[c..])
-                }
-            }
-            // if curr_len > 0 {
-            //     let c = curr_offset;
-            //     let res = if c > 0 {
-            //         curr_offset -= s.len();
-            //         if c < s.len()  {
-            //             &s[c..]
-            //         } else {
-            //             ""
-            //         }
-            //     } else {
-            //         s
-            //     };
-            //     Some(res)
-            })
-         .skip_while(|&s| s == "")
-    }
-
-    pub fn chars(&'a self) -> impl Iterator<Item=char> +'a  {
-        self.slice_char_iter(self.node.chars())
-    }
-
-    pub fn bytes(&'a self) -> impl Iterator<Item=u8> + 'a  {
-        self.slice_char_iter(self.node.bytes())
-    }
-
-    #[inline]
-    pub fn split_whitespace(&'a self) -> impl Iterator<Item=&'a str> {
-        self.slice_strings_iter(self.node.split_whitespace())
-    }
-
-    #[inline]
-    pub fn lines(&'a self) -> impl Iterator<Item=&'a str> {
-        self.slice_strings_iter(self.node.lines())
-    }
-
-    #[inline]
-    pub fn char_indices(&'a self) -> impl Iterator<Item=(usize, char)> + 'a {
-        self.chars().enumerate()
-    }
-
-    /// Returns true if the bytes in `self` equal the bytes in `other`
-    #[inline]
-    fn bytes_eq<I>(&self, other: I) -> bool
-    where I: Iterator<Item=u8> {
-        self.bytes().zip(other).all(|(a, b)| a == b)
-    }
-
-    #[inline]
-    pub fn len(&self) -> usize { self.len }
-}
-
-//-- comparisons ----------------------------------------------------
-impl<'a> cmp::Eq for RopeSlice<'a> {}
-impl<'a> cmp::PartialEq for RopeSlice<'a> {
-    /// A rope equals another rope if all the bytes in both are equal.
-
-    #[inline]
-    fn eq(&self, other: &RopeSlice<'a>) -> bool {
-        if self.len() == other.len() {
-            self.bytes_eq(other.bytes())
-        } else {
-            false
-        }
-    }
-}
-
-impl<'a> cmp::PartialEq<str> for RopeSlice<'a> {
-    /// A rope equals another rope if all the bytes in both are equal.
-
-    #[inline]
-    fn eq(&self, other: &str) -> bool {
-        if self.len() == other.len() {
-            self.bytes_eq(other.bytes())
-        } else {
-            false
-        }
     }
 }
