+++ an-rope/src/test.rs
@@ -276,61 +276,3 @@ fn rope_char_indices() {
         assert_eq!(ridx, sidx);
     }
 }
-
-#[test]
-fn rope_slice_char_indices() {
-    let string = "aaaaabbbbbbccccccccccccdefgdefgaabababab";
-    let rope = Rope::from(string);
-    let rope_slice = rope.slice(4..8);
-    let string_slice = &string[4..8];
-    let indices = rope_slice.char_indices().zip(string_slice.char_indices());
-    for ((ridx, rch), (sidx, sch)) in indices {
-        assert_eq!(rch, sch);
-        assert_eq!(ridx, sidx);
-    }
-}
-
-#[test]
-fn rope_slice_to() {
-    let string = "aaaaabbbbbbccccccccccccdefgdefgaabababab";
-    let rope = Rope::from(string);
-    let rope_slice = rope.slice(1..10);
-    let string_slice = &string[1..10];
-    assert_eq!(&rope_slice, string_slice)
-}
-
-// #[test]
-// fn rope_slice_between() {
-//     let string = "aaaaabbbbbbccccccccccccdefgdefgaabababab";
-//     let rope = Rope::from(string);
-//     let rope_slice = rope.slice(1...10usize);
-//     let string_slice = &string[1...10];
-//     assert_eq!(&rope_slice, string_slice)
-// }
-
-#[test]
-fn rope_slice_until() {
-    let string = "aaaaabbbbbbccccccccccdefgdefgaabababab";
-    let rope = Rope::from(string);
-    let rope_slice = rope.slice(..10);
-    let string_slice = &string[..10];
-    assert_eq!(&rope_slice, string_slice)
-}
-
-#[test]
-fn rope_slice_from() {
-    let string = "aaaaabbbbbbccccccccccccdefgdefgaabababab";
-    let rope = Rope::from(string);
-    let rope_slice = rope.slice(5..);
-    let string_slice = &string[5..];
-    assert_eq!(&rope_slice, string_slice)
-}
-
-#[test]
-fn rope_slice_full() {
-    let string = "aaaaabbbbbbccccccccccccdefgdefgaabababab";
-    let rope = Rope::from(string);
-    let rope_slice = rope.slice(..);
-    let string_slice = &string[..];
-    assert_eq!(&rope_slice, string_slice)
-}
